refactor(api): consolidate SSE event bus and Kubernetes polling
